feat: enhance subscription features and secure access control
Refactor the ` UserSubscription ` entity and improve attribute definitions. Add lazy loading to related entities. Update `SubscriptionController` to include user-specific subscription actions with authentication. Implement access control checks for subscription cancellation.

+++ streaming-service/src/main/java/dev/studentpp1/streamingservice/subscription/dto/UserSubscriptionDto.java
@@ -4,9 +4,11 @@ import dev.studentpp1.streamingservice.subscription.entity.SubscriptionStatus;
 import java.time.LocalDateTime;
 
 public record UserSubscriptionDto(
-        Integer id,
-        LocalDateTime startTime,
-        LocalDateTime endTime,
-        SubscriptionStatus status,
-        String planName) {
+    Long id,
+    LocalDateTime startTime,
+    LocalDateTime endTime,
+    SubscriptionStatus status,
+    String planName
+) {
+
 }
